refactor: header now is single component + has animation
