--- src/Home/Header/Logo/index.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-import logo from "./logo.svg";
-
-export default function Logo(props) {
-  return <img src={logo} alt={props.alt} />;
-}
