--- src/Search/Header/styles.sass
@@ -1,68 +1,0 @@
-@import "../../global.sass"
-.search-header
-  padding-top: 12px
-  padding-bottom: 32px
-  background: linear-gradient(163.71deg, #00B0DE 0.36%, #01AEDC 4.04%, #02ABDB 7.77%, #02ABDB 11.48%, #02ABDB 11.7%, #196EBD 100%)
-  &__wrapper
-    position: relative
-  &__set-currency
-    position: absolute
-    top: 8px
-    right: 8px
-    font-family: Roboto
-    font-weight: 500
-    font-size: 16px
-    text-align: center
-    color: #FFFFFF
-    background: rgba(0, 0, 0, 0.0001)
-    border: 1px solid rgba(255, 255, 255, 0.5)
-    border-radius: 100px
-    padding: 8px 20px 8px 20px
-    +desktop
-      right: 0
-  &__logo
-    display: flex
-    align-items: center
-    margin-bottom: 40px
-    text-decoration: none
-    &__title
-      font-family: Roboto
-      line-height: 25px
-      font-size: 20px
-      color: #FFFFFF
-      margin-left: 12px    
-  &__search-button
-    font-family: Roboto
-    font-weight: 500
-    font-size: 20px
-    color: #FFFFFF    
-    background: #FF9241
-    padding: 16px 24px 16px 24px
-    border: none
-    flex-basis: 52%
-    margin-left: 2px
-    height: 54px
-    border-bottom-right-radius: 4px
-    white-space: nowrap
-    cursor: pointer
-    +desktop
-      margin-left: 16px
-      border-radius: 4px
-      margin-top: 2px
-
-.fixer
-  flex-basis: 117%
-.header-search__search-filters__passengers-picker
-  background: #FFFFFF 
-  padding: 16px
-  margin: 2px 0
-  +tablet
-    flex-basis: 50%
-    margin-left: 2px
-    margin-top: 0
-    border-bottom-right-radius: 0px
-  +desktop
-    margin-top: 2px
-    flex-basis: 50%
-    border-top-right-radius: 4px
-    border-bottom-right-radius: 4px
